==> r_docs/01_load.R <==

library(plyr)
library(tidyverse)
library(magrittr)
library(purrr)
library(flextable)
library(gtsummary)
library(survival)
library(survminer)
library(rms)
library(mice)
library(sjmisc)
library(cmprsk)
library(prodlim)
library(lubridate)
library(ftExtra)
library(rstpm2)
library(cowplot)
library(forestplot)
library(ftExtra)
library(huxtable)
library(grid)
library(gridExtra)
library(webshot2)
library(forcats)
library(magick)
library(tidycmprsk)
library(hardhat)
library(interactionRCS)
